Correct attribute list parsing
The previous attribute list parsing implementation split tags on the '='
character even when the character was part of a quoted string. This was
problematic when working with attribute values containing a '='
character (e.g. a URL with parameters).

This change replaces the previous regex-based implementation with a
tokenization-based solution that is aware of quoted strings. Note that
this implementation will improperly accept illegal attribute values that
contain a backslash-escaped double quote (\") and interpret it as a
literal double quote; however, this is unlikely to pose a problem in
practice.